Add files via upload
Q3 of State Exercises : Add another piece of state: the room temperature. Initialize it to 72 (or 22 for you Celsius types). Display the current temperature inside render() .  Add 2 more buttons: “+” and “-”. Then add corresponding functions that will increase or decrease the temperature by 1 degree when clicked, and wire up the buttons.